+++ jdaviz/configs/default/plugins/viewers.py
@@ -4,8 +4,18 @@ from glue_jupyter.table import TableViewer
 
 from jdaviz.core.registries import viewer_registry
 
-__all__ = []
+__all__ = ['JdavizViewerMixin']
 
 viewer_registry.add("g-profile-viewer", label="Profile 1D", cls=BqplotProfileView)
 viewer_registry.add("g-image-viewer", label="Image 2D", cls=BqplotImageView)
 viewer_registry.add("g-table-viewer", label="Table", cls=TableViewer)
+
+
+class JdavizViewerMixin:
+    @property
+    def _jdaviz_app(self):
+        return self.session.jdaviz_app
+
+    @property
+    def _jdaviz_helper(self):
+        return self._jdaviz_app._jdaviz_helper