+++ CHANGES.rst
@@ -32,6 +32,9 @@ API Changes
 - Viewers now can access the calling Jdaviz application using
   ``viewer.session.jdaviz_app``. [#1051]
 
+- Viewers now can access the Jdaviz application via ``viewer._jdaviz_app`` and the
+  helper via ``viewer._jdaviz_helper``. [#1054]
+
 Cubeviz
 ^^^^^^^